Remove DecoratorPanel.java from source.  TextItemPanel now handles all text modification. This removes an awkward dialog from Editor.
