+++ java/src/jmri/jmrit/display/palette/ClockItemPanel.java
@@ -48,6 +48,7 @@ public class ClockItemPanel extends IconItemPanel {
             add(Box.createVerticalGlue());
         }
         _previewPanel.invalidate();
+        hideIcons();
     }
 
     @Override
