+++ java/src/jmri/jmrit/display/palette/DragJComponent.java
@@ -13,6 +13,8 @@ import java.awt.dnd.DragSourceDropEvent;
 import java.awt.dnd.DragSourceEvent;
 import java.awt.dnd.DragSourceListener;
 import javax.swing.BorderFactory;
+import javax.swing.Box;
+import javax.swing.BoxLayout;
 import javax.swing.JComponent;
 import javax.swing.JPanel;
 import org.slf4j.Logger;
@@ -37,17 +39,16 @@ import org.slf4j.LoggerFactory;
 public abstract class DragJComponent extends JPanel implements DragGestureListener, DragSourceListener, Transferable {
 
     DataFlavor _dataFlavor;
+    JComponent _component;
 
     public DragJComponent(DataFlavor flavor,  JComponent comp) {
         super();
+        _component = comp;
         setBorder(BorderFactory.createTitledBorder(BorderFactory.createLineBorder(Color.black),
                 Bundle.getMessage("dragToPanel")));
         // guestimate border is about 5 pixels thick. plus some margin
         add(comp);
-        Dimension dim = comp.getPreferredSize();
-        int width = Math.max(100, dim.width + 20);
-        int height = Math.max(65, dim.height + 20);
-        setPreferredSize(new java.awt.Dimension(width, height));
+        setPreferredSize();
         setToolTipText(Bundle.getMessage("ToolTipDragIcon"));
         DragSource dragSource = DragSource.getDefaultDragSource();
         dragSource.createDefaultDragGestureRecognizer(this,
@@ -55,6 +56,23 @@ public abstract class DragJComponent extends JPanel implements DragGestureListen
         _dataFlavor = flavor;
     }
 
+    @Override
+    public void invalidate() {
+        setPreferredSize();
+        super.invalidate();
+    }
+
+    protected void setPreferredSize() {
+        Dimension dim = _component.getSize();
+        int width = Math.max(100, dim.width + 30);
+        int height = Math.max(65, dim.height + 35);
+        setPreferredSize(new java.awt.Dimension(width, height));
+    }
+
+    protected JComponent getcomponent() {
+        return _component;
+    }
+        
     protected boolean okToDrag() {
         return true;
     }
@@ -84,22 +102,18 @@ public abstract class DragJComponent extends JPanel implements DragGestureListen
 
     @Override
     public void dragEnter(DragSourceDragEvent e) {
-        //if (log.isDebugEnabled()) log.debug("DragJLabel.DragSourceDragEvent ");
     }
 
     @Override
     public void dragExit(DragSourceEvent e) {
-        //if (log.isDebugEnabled()) log.debug("DragJLabel.dragExit ");
     }
 
     @Override
     public void dragOver(DragSourceDragEvent e) {
-        //if (log.isDebugEnabled()) log.debug("DragJLabel.dragOver ");
     }
 
     @Override
     public void dropActionChanged(DragSourceDragEvent e) {
-        //if (log.isDebugEnabled()) log.debug("DragJLabel.dropActionChanged ");
     }
 
     /**
@@ -107,13 +121,11 @@ public abstract class DragJComponent extends JPanel implements DragGestureListen
      */
     @Override
     public DataFlavor[] getTransferDataFlavors() {
-        //if (log.isDebugEnabled()) log.debug("DragJLabel.getTransferDataFlavors ");
         return new DataFlavor[]{_dataFlavor, DataFlavor.stringFlavor};
     }
 
     @Override
     public boolean isDataFlavorSupported(DataFlavor flavor) {
-        //if (log.isDebugEnabled()) log.debug("DragJLabel.isDataFlavorSupported ");
         return _dataFlavor.equals(flavor);
     }
